Some ids are now classes. Force stuff below portrait.
dex-page-portrait, dex-page-beside-portrait, and dex-page-damage are now
classes.  I needed more than one dex-page-damage on type pages.  I was
changing the others anyway and thought they should be classes, but it now
occurs to me that it doesn't actually make a difference yet.  But it might in
the future, I guess, and there's nothing gained by having them be ids.

dex-page-beside-portrait:after is now clear:both; this way, if there's little
enough content in dex-page-beside-portrait that the next thing would
otherwise slip beside the portrait, it won't.  Turns out this is never
actually a problem; it wasn't before, but I thought it might be on type
pages.  Good thing it's not, because apparently IE7 and before don't support
:after.