Sorted dex-pokemon-moves tables into colgroups and tbodies.
I didn't know whether to split section headers into their own tbodies
or not; iirc, I chose to.  Also, I was iffy on not renaming move_columns
given that it's not a list of move column *groups*, but it's also, in
the end, primarily a list of move columns—they're only grouped by
generation to make forming colgroups easier.

Signed-off-by: Eevee <git@veekun.com>