Prettify Pokémon form explanation display.
I didn't know we already had a markdown class that included list
bullets!  Also, they gain their own paragraph tags, which was resulting
in empty paragraphs, so I applied the class to a div instead.

I removed them outright from flavour pages because... they're not
flavour.  The actual form selector is still there, but the explanations
are game mechanics, and a few of them have reached multiple paragraphs
in length.