Make caching play nicely with obdurate.
The obdurate cheat code subtly alters the way flavor text is rendered.
Most pages are cached. Caching doesn't take the effect of cheat codes
into account. Cheat codes are therefore very flaky.

This is a long-standing bug.

The fix is simply to add the cheat code state to the cache key. An
alternative would be to bypass the cache when cheat codes are enabled.

Caveat: this commit causes location and type pages, which don't have
any flavor text, to have redundant copies cached.