Replace eagerload with joinedload.
Joinedload is a synonym for eagerload since SQLAlchemy 0.6.

Although eagerload isn't deprecated and it looks like it will be
around for the forseeable future, there's no good reason to hang onto
it, and there's *really* no good reason to be using both names. Let's
pick one and stick with it.

Fixes #61.