Tables => t
Import pokedex.db.tables under a much shorter name. Not sure where this
convention started, but it has been in the air for quite a while.
Time to bring it to spline-pokedex.

Causes a lot of churn so of course nothing will merge cleanly across
this. An acceptable price.

Zhorken says, "hell yes".