Move image and link helpers to dexlib
Pyramid requires access to the request object to be able to generate
URLs, so move helpers that generate URLs to dexlib, where they'll have
access to the template context. This seems easier than adding an
explicit 'request' or 'url' parameter to all of them.

The next commit will fix up calls to the old functions.

- The _=_ parameters are no longer necessary because we already have
  access to the translator. Keep the parameters so that we don't break
  callers, but change the default to _=None because there is no longer
  a global _ variable in scope.

Updates #115.
Updates #117.