Add encounter icons for gift & gift-egg methods
This commit adds icons for the soon-to-be-added gift and gift-egg
encounter methods (veekun/pokedex#238).

The gift icon is from the attack animation for Present in Gen IV
and was manually cropped from
https://cdn.bulbagarden.net/upload/d/de/Present_IV_heal.png.
(I had to shrink it down a bit to get it to fit with the other encounter
method icons; gift-big.png is the original size.)

The egg icon is from pokemon/icons/egg.png in pokedex-media.