Add explicit config param to pokemon_has_media
For the pyramid port, which doesn't have threadlocal globals like
config. Most of the other helpers are going to be moving to dexlib,
where they'll have access to the request context, but pokemon_has_media
has to stay because it is used in a controller, so we have to get the
context to it some other way.

Also add a convenient wrapper function to dexlib which passes the config
parameter implicitly, and update the template code to call it.

Updates #117