Pokémon search: escape pokedex labels
WTForms 2.3 started escaping label text with markupsafe[1].
We have several cases where we actually do want the label to contain
html (to show icons). In most cases we were already passing an object
with an __html__ method so nothing needs to be changed. The only
exception is the list of regional pokedexes on the pokemon search page.

[1]: https://github.com/wtforms/wtforms/issues/315

Updates #130